# db: reduce allocations in range key iteration